Returning to the current service after registration; slight refactoring
