+++ src/api/config.ts
@@ -1,2 +1,2 @@
-export const baseUrl = "https://cashwizkiosk.info:8080/api"; // PROD
-// export const baseUrl = "http://localhost:8080/api"; // DEV
+// export const baseUrl = "https://cashwizkiosk.info:8080/api"; // PROD
+export const baseUrl = "http://localhost:8080/api"; // DEV
